+++ openrem/remapp/urls.py
@@ -119,7 +119,10 @@ urlpatterns += patterns('remapp.exports.exportviews',
     url(r'^download/(?P<task_id>[a-f0-9-]{36})$', 'download'),
     url(r'^deletefile/$', 'deletefile'),
     url(r'^export/abort/(?P<pk>\d+)$', 'export_abort'),
-)
+    url(r'^export/updateactive$', 'update_active', name='update_active'),
+    url(r'^export/updateerror$', 'update_error', name='update_error'),
+    url(r'^export/updatecomplete$', 'update_complete', name='update_complete'),
+    )
 
 urlpatterns += patterns('remapp.exports',
                         url(r'^xlsx/openrem/ct/',