+++ docs/changes.rst
@@ -4,6 +4,7 @@ OpenREM version history
 
 0.8.0 (?)
 ---------
+* `#575`_  Exports: export page now updates using AJAX and has a select all button
 * `#573`_  Exports: corrected and clarified exposure time and duration units, added number of pulses
 * `#572`_  Interface: Homepage now populates as AJAX to increase responsiveness
 * `#566`_  Interface: Correctly show "assumed patient mass" in case of set value of zero